feat(sharepoint-improvements): Sharepoint improvements (#73)
## Describe your changes
- `list-shared-sites` renamed from `list-sites` an action to list the
user's shared sites. It no longer writes to the metadata but rather
leaves that up to the end user. The metadata is now just an array of
sharepoint sites
- `shared-sites-selection` renamed from `file-sync` sites the files from
the PickedSites which as mentioned is now just an array of strings
(which are sites)
- `user-files` a new sync that iterates over all the user's drives and
recursively fetches all files and saves them
- `user-files-selection` new sync that simply syncs files that are set
in the metadata as selected by the user

## Issue ticket number and link
NAN-1914

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is:
- [ ] External API requests have `retries`
- [ ] Pagination is used where appropriate
- [ ] The built in `nango.paginate` call is used instead of a `while
(true)` loop
- [ ] Third party requests are NOT parallelized (this can cause issues
with rate limits)
- [ ] If a sync requires metadata the `nango.yaml` has `auto_start:
false`
- [ ] If the sync is a `full` sync then `track_deletes: true` is set