feat(ring-central): add RingCentral users integration  (#78)
## Describe your changes
Add RingCentral integration with the following endpoints:

### Syncs
- users

### Actions
- create-user
- delete-user

**Important Note: No fixtures, nor tests/mocks were generated. I could
not test any of the syncs and actions added to the RingCentral
integration as it requires an enterprise account, which is not
accessible to me. Hence, once a connection is setup in Nango, these
endpoints need to be tested and validated**

## Issue ticket number and link
N/A

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: No fixtures, nor
tests/mocks were generated. I could not test any of the syncs and
actions added to the RingCentral integration it requires an enterprise
account, which is not accessible to me.
- [X] External API requests have `retries`
- [X] Pagination is used where appropriate
- [X] The built in `nango.paginate` call is used instead of a `while
(true)` loop
- [ ] Third party requests are NOT parallelized (this can cause issues
with rate limits)
- [ ] If a sync requires metadata the `nango.yaml` has `auto_start:
false`
- [X] If the sync is a `full` sync then `track_deletes: true` is set

---------

Co-authored-by: Khaliq <khaliqgant@gmail.com>
Co-authored-by: Khaliq <khaliq@nango.dev>