fix(docs): Fix issue relating to symlinked configs (#171)
## Describe your changes

The issue here was that I didn't realize we used symlinks within the
repo. We would iterate the non-sandbox version of the docs and generate
the readme, and then iterate the sandbox version and generate the readme
through a symlink, which replaced the non-sandbox one.

This updates the script to ignore any symlinked integrations.

## Issue ticket number and link

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)

-   [ ] I added tests, otherwise the reason is:
-   [ ] External API requests have `retries`
-   [ ] Pagination is used where appropriate
- [ ] The built in `nango.paginate` call is used instead of a `while
(true)` loop
- [ ] Third party requests are NOT parallelized (this can cause issues
with rate limits)
- [ ] If a sync requires metadata the `nango.yaml` has `auto_start:
false`
-   [ ] If the sync is a `full` sync then `track_deletes: true` is set
- [ ] I followed the best practices and guidelines from the [Writing
Integration
Scripts](/NangoHQ/integration-templates/blob/main/WRITING_INTEGRATION_SCRIPTS.md)
doc